Improved OutputManager class and project structure
- Renamed resources.py to system.py and adjusted all references and imports across the modules
- Improved OutputManager class in system.py and made it more robust
- Moved MODEL_METADATA (previously just called MODELS) to models.py
- Moved cropped_fov_prediction_pipeline() from image_processing.py to predict.py
- Added some checks for more safety across all modules
- predict_from_array_by_iterator() now needs the accelerator as str. It cant be determined automatically anymore.
- Adjusted doc-strings for a few functions
